Separating the Pub Sub functionalities in a different class
